Fix double deletes
When running multiple test workers against the same account it’s possible that two of them try to delete the same app, when this happens we should not be raising an error, but instead see if the user is under the “threshold”, if they are do nothing, if not the next app should be removed. Repeat until the user is under the threshold.