Introduce Lock file to Hatchet
Currently the state of a Project’s tests using Hatchet is dependent on the shared global state of the test app source code. Unfortunately this is mutable global state. This means that without anything changing in a project that the test suite can suddenly start failing if an external app is mutated.

To address this issue, this PR introduces the concept of a “lock” file. It works like this, when you run your `hatchet install` the git sha of each repo is checked against a previously recorded version in the `hatchet.lock`. 

There some differences to bundler. The `Gemfile` the `hatchet.json` cannot specify versions or ranges. Therefore the only source of versioning comes in the `hatchet.lock`

The command `$ hatchet install` currently does not update the `hatchet.lock`. If you want to add a dependency add it to your `hatchet.json` first and then run `$ hatchet lock`. If you want to update a repo, either `git pull` it locally and then run `hatchet lock`, or you can manually change the sha in the `hatchet.lock` and then run `hatchet install`.