Add API rate limits
When deploying lots of apps for a buildpack it’s very easy to hit the Heroku API rate limit. When this happens builds start to fail with 422 status. It’s pretty cryptic. Usually what happened is there are multiple CI services running at once and they flooded the API.

This patch adds a tiny sleep before each API call. The sleep is proportional to the amount of capacity left in your rate limit bucket. If you have lots of requests left then the sleep is barely noticeable. If you have only 1 left, then you’ll be sleeping for a full 60 seconds.