Fix race condition in Reaper#cycle
Concurrent deletions of the same app could cause a race condition in API, resulting in 403s that caused tests to fail.

We are now using a Mutex to prevent multiple processes (when using parallel_tests) from potentially deleting the same app at the same time.

The method has been refactored to fetch the list of apps only once, and use iteration rather than recursion, to work through the list.

The fetching-once is much more efficient, as it decreases the number of API calls drastically.

Recursion can no longer be used as the innermost call exiting would otherwise lift the file lock in its ensure block.

As other processes will be waiting for a lock inside cycle, or soon hitting cycle after a deploy block end, or as a result of App#create failing due to an exceeded limit, we can rely on that instead of re-fetching the list of apps all the time to clean up apps that may have been created while we were cycling.