Ensure mutex file is closed when there is an error
The mutex file in the reaper is used to ensure that only one process
will try to destroy apps given a multi process system. This is
accomplished via an `ensure` block. However this `ensure` block only
fires when the context is exited, and in the case where `retry` fires,
`ensure` is skipped.

This PR attempts to avoid deadlock by ensuring that the mutex file is
closed even when `retry` is called.