Remove bundle_exec command
This bundler command is being deprecated:

```
[DEPRECATED] `Bundler.with_clean_env` has been deprecated in favor of `Bundler.with_unbundled_env`. If you instead want the environment before bundler was originally loaded, use `Bundler.with_original_env` (called at /home/travis/build/heroku/hatchet/lib/hatchet/app.rb:418)
```

We're using it in places where we were shelling out to `heroku`. I believe this was because the `heroku` cli used to be distributed as a ruby gem and in case the user did not have it in their Gemfile then the command would fail. 

The `heroku` command hasn't been distributed as a gem for a long time so I believe we can safely remove this, but the tests and time will tell.