Allow for Test re-runs
I needed to run a test twice to assert that the cache was used for the Ruby buildpack. I hacked it:

https://github.com/heroku/heroku-buildpack-ruby/blob/facd23708feff15d5bdbfd0455dc470a21b11a76/spec/hatchet/ci_spec.rb#L41-L43

This PR gives a supported API instead to call `TestRun#run_again`.

While working on this, I discovered that the test run wasn't being run inside the temp directory. So I fixed that bug, I also added support for `before_deploy`.