Bundle context is needed to find `hatchet` executable
Hatchet is installed as a gem, without `bundle exec` it won't be found. The directive `Bundler.with_clean_env` is how the environment is set up before `bundle install`. I don't know how this ever worked before unless travis was putting hatchet on the path or `bin` was already in the path somehow.