Fail on release failure
When a release fails to execute the app is never deployed, for example if you were to run:

```ruby
git push heroku master && heroku run ls
```

Both commands will run, but `heroku run ls` will be empty since the release never finished. This creates weird to debug hatchet tests as the deploy will work but any subsequent `app.run(<cmd>)` will fail in strange ways.

This PR seeks to make this behavior more explicit, by explicitly failing on an incomplete release. This is a breaking change so it's a major version bump.