[close #89] Set minimum Time to Live for apps
There are cases where Hatchet deletes an app as it is being used/deployed. For example if you've configured a test runner to concurrently execute 25 tests, but you've set HATCHET_APP_LIMIT=20` then one test process will create an app as another is trying to destroy it. This results in odd errors like this in the deploy:

```
       remote: !	No such app as hatchet-t-522c335bea.
```

To prevent this we can set a configurable minimum time to live for applications. In the above scenario where we want to run 25 apps, but have set a maximum of 20 then Hatchet would wait until the oldest app was 7 minutes old before deleting it. This should give the test a chance to finish before continuing and deleting the application.

This value can be configured by the environment variable:

```
HATCHET_MINIMUM_TTL_MINUTES=7
```

When this behavior is triggered a warning is emitted so the developer can investigate and figure out how to re-configure/tune the system so that it doesn't continue to happen.