Remove support for passing a block to App#run (#105)
Since it's been deprecated since Hatchet 5.0.0, and removing support
means we can drop the dependency on `repl_runner`, and so one of the
consumers of `activesupport`.

The addition of `activesupport` to Hatchet's gemspec is required since
Hatchet itself relies upon `activesupport` but was previously relying on
it being pulled in implicitly via `repl_runner`. Later PRs will remove
the dependency on `activesupport` entirely.

Fixes part of #103.