[close #85] Allow `hatchet lock` to work on new repo
Currently when you run `hatchet install` for the first time on a new app it tells you to first run `hatchet lock`. However when you run `hatchet lock, it fails due to a missing file. 

This patch tests that behavior and allows for running `hatchet lock` when the file does not exist