[close #117] Allow different branch than default
If you've locked to the `main` branch in a lockfile, but the default branch on GitHub is something different (like `master`) then when hatchet tries to run `git reset --hard main` it fails because it doesn't have the references for `main` yet.

This commit fixes this by first explicitly fetching that commit, checking out into it to make sure it exists, and then doing the reset.