Fixes to `hatchet lock` (#118)
* [close #85] Allow `hatchet lock` to work on new repo

Currently when you run `hatchet install` for the first time on a new app it tells you to first run `hatchet lock`. However when you run `hatchet lock, it fails due to a missing file. 

This patch tests that behavior and allows for running `hatchet lock` when the file does not exist

* [close #117] Allow different branch than default

If you've locked to the `main` branch in a lockfile, but the default branch on GitHub is something different (like `master`) then when hatchet tries to run `git reset --hard main` it fails because it doesn't have the references for `main` yet.

This commit fixes this by first explicitly fetching that commit, checking out into it to make sure it exists, and then doing the reset.