[close #109] Allow retrying failed release (#119)
* [close #109] Allow retrying failed release

When a deploy succeeds but the release fails the git remote retains the contents of the `git push`. This means if you immediately try to push again that you'll get a `Everything up-to-date` from the remote and Heroku will not try to re-build your project.

We don't want that behavior, we want to retry on a failed release incase it's intermittent. This PR adds an empty commit when a failed release is detected.

* Test `allow_failure` retry count behavior

When `allow_failure` is set to true then it over-writes the value set by `retries`, this test asserts that to be true.