Fix support for deploying `main` branch
When I shipped support for the main branch I only checked that the lock file correctly installed the main branch, but never checked it could be deployed. 

In our deploy we previously ran `git push heroku master` which is short for `git push master:master` and this fails since there's no master branch of that repo. This force deploys HEAD instead.