[close #83] HATCHET_BUILDPACK_BASE updates (#133)
Currently if you want to set HATCHET_BUILDPACK_BASE, you need to do so before you `require 'hatchet'` which is unexpected. By evaluating this env var every time a new app is created it allows the developer to set this env var when they want instead of before requiring the library.

The other change here is to no longer default to a buildpack. While it was useful for defaulting to the Ruby buildpack when only the ruby buildpack used this library, it is confusing when setting up a new buildpack to have deploys seemingly happen but it to not use your buildpack code. This can be fixed by raising a helpful error when HATCHET_BUILDPACK_BASE is not set. However that's a breaking change. For now I'm just deprecating that fallback behavior.