[close #50] Annotate rspec errors (#136)
For intermittent failures it's important to get as much debugging information as possible about the test in progress. This PR annotates any rspec expectation failures that happen in a `deploy {}` block with the application name. Here's an example of what it will look like:

```
     Failure/Error: expect(true).to eq(false)

       App: hatchet-t-873bcea87c (default_ruby)

       expected: false
            got: true

       (compared using ==)

       Diff:
       @@ -1 +1 @@
       -false
       +true
``` 


In this example if the failure was due to an `app.run` then we could manually `app.run` against the same app (assuming it had not been reaped yet) in order to better debug.