[close #156] Capture stderr from `heroku run` (#187)
* [close #156] Capture stderr from `heroku run`

The rate limiting messages are delivered via the `heroku` cli to stderr which we're currently not capturing. This means that when we check stdout there will never be a rate limit message, so our rate limiting and retry code never fires. This is documented in #156.

This PR fixes this by moving from exec-ing the process via backpacks to using `Open3.capture3` which captures stderr in addition to stdout. The biggest caveat is that it does not set the value for $? Which some tests are likely relying on. To prevent this change from breaking backwards compat we will exec an `exit` with the same status code. It's not ideal, but it will work.

In the future we may want to consider deprecating returning a string and moving to return an object that holds a status value instead.

* [close #188] Use non-deprecated stack in test

Heroku-16 is deprecated. Heroku-18 is not.

```
{"id":"unsupported","message":"Creating new apps on the heroku-16 stack is not supported. Please consider using heroku-20 instead: https://devcenter.heroku.com/articles/stack. If you need to create a new heroku-16 app, please open a support ticket: https://help.heroku.com/new/increase-limits."}
```