Rename on_conflict strategies
The original reason conflict strategies were added was to gate whether or not we hit the API again. All times a conflict is hit a sleep occurs, so in the strategy `:sleep_and_continue`, the least important part is the sleep. I renamed it to put the behavior of refreshing the API front and center.

Also updated bin/hatchet to make it explicit instead of relying on defaults.